refactor: update component preview styling and structure
- Remove gap-0 class from buttons container in ComponentPreviewV2
- Add aria-label to replay button in ComponentPreviewV2
- Replace ComponentPreviewV2 with ComponentPreview in mdx imports
- Update ComponentPreview to use base tabs with indicator
- Rename notProse prop to prose with inverted logic
- Simplify tab triggers by removing icons
- Update button styling from secondary to ghost variant
- Fix code content selector using modern CSS syntax
